requirements.txt
flask
pillow
scikit-learn
requests
